Move the Local Artifact Services to the DirectRunner
This is expected to be the only user, at the current time.
