--- runners/local-artifact-service-java/build.gradle
@@ -1,44 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * License); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an AS IS BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-apply from: project(":").file("build_rules.gradle")
-applyJavaNature()
-
-description = "Apache Beam :: Runners :: Java Local Artifact Service"
-ext.summary = """The Beam Artifact Service exposes APIs to stage and retrieve
-artifacts in a manner independent of the underlying storage system, for use
-by the Beam portability framework. The local implementation uses the local
-File System as the underlying storage system."""
-
-dependencies {
-  compile library.java.guava
-  shadow project(path: ":beam-model-job-management", configuration: "shadow")
-  shadow project(path: ":beam-model-pipeline", configuration: "shadow")
-  shadow project(path: ":beam-runners-java-fn-execution", configuration: "shadow")
-  shadow library.java.findbugs_jsr305
-  shadow library.java.grpc_core
-  shadow library.java.grpc_stub
-  shadow library.java.protobuf_java
-  shadow library.java.slf4j_api
-  testCompile project(path: ":beam-runners-core-construction-java")
-  testCompile library.java.hamcrest_core
-  testCompile library.java.hamcrest_library
-  testCompile library.java.mockito_core
-  testCompile library.java.junit
-  testCompile library.java.slf4j_jdk14
-}
